requirements.txt
certifi
charset-normalizer
clipboard
darkdetect
hjson
idna
llvmlite
numba
numpy
pillow
pygame
pynput
pyperclip
PyQt5
PyQt5-Qt5
PyQt5_sip
pyqtdarktheme
QScintilla
requests
six
urllib3
pyinstaller
orjson
cython
ujson
setuptools
matplotlib
pypresence
orjson
